Moving CompilerContext into context directory
