+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/TruffleCompilerContext.java
@@ -10,10 +10,10 @@ import java.util.function.Consumer;
 import java.util.logging.Level;
 import org.enso.compiler.Cache;
 import org.enso.compiler.Compiler;
-import org.enso.compiler.CompilerContext;
 import org.enso.compiler.PackageRepository;
 import org.enso.compiler.SerializationManager;
 import org.enso.compiler.codegen.IrToTruffle;
+import org.enso.compiler.context.CompilerContext;
 import org.enso.compiler.context.InlineContext;
 import org.enso.compiler.core.IR;
 import org.enso.compiler.data.CompilerConfig;
