+++ engine/runtime/src/main/java/org/enso/interpreter/runtime/builtin/Builtins.java
@@ -17,7 +17,7 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
-import org.enso.compiler.CompilerContext;
+import org.enso.compiler.context.CompilerContext;
 import org.enso.compiler.Passes;
 import org.enso.compiler.context.FreshNameSupply;
 import org.enso.compiler.core.CompilerError;
